GCPmodel/requirements.txt
# Classical Literature AI - Requirements
# ===========================================

# Core Model & Training
transformers>=4.38.0
peft>=0.9.0
accelerate>=0.27.0
bitsandbytes>=0.42.0
datasets>=2.17.0
torch>=2.1.0

# Evaluation System
google-generativeai>=0.4.0
google-cloud-aiplatform>=1.40.0

# Language Analysis
konlpy>=0.6.0
sentence-transformers>=2.2.2
scikit-learn>=1.3.0

# Utilities
pandas>=2.1.0
numpy>=1.24.0
pyyaml>=6.0.0
tqdm>=4.65.0
python-dotenv>=1.0.0

# API Integration
openai>=1.0.0
httpx>=0.24.0

# Development & Testing
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
